Moved ByteArrayImpl to common
